tests, readmes, some logic fixed

+++ ( 2 ) second Year/( 3 )  Third semester/Data structures/Homework/src/Headers/Board.hpp
@@ -30,7 +30,7 @@ public:
     }
  
 
-    Pair<int> IndexToPair(int index) const;
+    Pair<short> IndexToPair(int index) const;
 
     void InitBoard(int rows = 8, int collums = 8);
